requirements.txt
langgraph>=0.4
langchain-openai>=0.3
langchain-tavily>=0.1
langchain-core>=0.3
pydantic>=2.0
python-dotenv>=1.0
pyyaml>=6.0
pytest>=8.0
